make activities dialogs more generic, and move workflow activity dialog logic to generic dialog
