+++ app/components/activities/dialogs/activity_list_dialog_component.html.erb
@@ -44,7 +44,7 @@
     <template id="listRow" data-activities--extended_details-target="listRow">
       <li class="pt-3 pb-3 sm:pb-4 sm:pt-4">
         <p class="inline-flex text-sm text-slate-500 truncate dark:text-slate-400">
-          <span class="mr-2" data-activities--extended_details-target="sampleName"></span>
+          <span class="mr-2" data-activities--extended_details-target="itemName"></span>
           <%= render PuidComponent.new(puid: "", show_clipboard: false) %>
         </p>
       </li>
