+++ app/components/activities/dialogs/activity_table_listing_dialog_component.html.erb
